refactor pagination e2e (#1312)
* wrap test suites in functions

* better naming and small improvements

* higher suite that calls the wrapped test functions
